in-progress imputation
